% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneland_parallel_FUNs.R
\name{adm_process}
\alias{adm_process}
\title{burning as burning/thinning}
\usage{
adm_process(path.mcmc.noadm, adm, nrun, burnin, coord, pal = "Set1")
}
\description{
burning as burning/thinning
}
