% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneland_parallel_FUNs.R
\name{make.folders}
\alias{make.folders}
\title{Create folders for each run in output}
\usage{
make.folders(main, nrun)
}
\arguments{
\item{main}{Root directory}

\item{nrun}{Number of runs}
}
\description{
Used internally, but can be used also to re-create the paths to results
}
