% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneland_parallel_FUNs.R
\name{make.npop.plots}
\alias{make.npop.plots}
\title{Make a plot of the sampled number of clusters during the chain}
\usage{
make.npop.plots(d, bur = burnin, th = nthin)
}
\description{
Make a plot of the sampled number of clusters during the chain
}
