% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotProbMembership_FUN.R
\name{PlotProbMembership}
\alias{PlotProbMembership}
\title{Generate a bar plot with probability of membership for each sample}
\usage{
PlotProbMembership(path.mcmc, pal = "Set1", metadata, orderBy)
}
\arguments{
\item{path.mcmc}{Path where the outputs are stored}

\item{pal}{Palette to use to colour code membership}

\item{orderBy}{Optional: name of the column in metadata to order the samples by}

\item{matadata}{Optional: data.frame with nindiv lines with metadata associated with samples 
as columns. One of these can be used to sort the samples}
}
\description{
Generate a bar plot with probability of membership for each sample
}
