% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneland_parallel_FUNs.R
\name{run_paral_geneland}
\alias{run_paral_geneland}
\title{Run Geneland in parallel
nrun Number of parallel runs
ncores Number of cores to use. If default (NULL) are automatically selected
model Whether "Correlated" or "Uncorrelated"
path Path to root directory where to store results
main.dir Directory name where to store output (within path)
spatial Whether to use the spatial model
GEN Input genotypes (for now only codominant are available)
COORD Coordinates
jitter Spatial uncertainty
nxdom,nydom Number of pixel for discretization of the spatial domain in the 
        horizontal and vertical direction respectively. If "auto" is automatically set
        otherwise a numeric vector of length=1
burnin Burnin
POPSMAX Maximum number of pops
POPINIT Initial number of pops
POPMIN Minimum number of pops
ITR Number of iteration
THN Thinning
nullMatrix A matrix with nindiv lines and nloc columns of 0 or 1. For each 
         individual, at each locus it says if the locus is genuinely missing 
         (no attempt to measure it). This info is used under the option 
         filterNA=TRUE do decide how a double missing value should be treated 
         (genuine missing data or double null allele).}
\usage{
run_paral_geneland(
  nrun,
  ncores = NULL,
  model = "Correlated",
  main.dir = "output_corr",
  spatial = TRUE,
  path,
  GEN,
  COORD,
  jitter = 0,
  nxdom = "auto",
  nydom = "auto",
  burnin,
  POPSMAX,
  POPINIT,
  POPMIN,
  ITR,
  THN,
  nullMatrix = NULL
)
}
\description{
Run Geneland in parallel
nrun Number of parallel runs
ncores Number of cores to use. If default (NULL) are automatically selected
model Whether "Correlated" or "Uncorrelated"
path Path to root directory where to store results
main.dir Directory name where to store output (within path)
spatial Whether to use the spatial model
GEN Input genotypes (for now only codominant are available)
COORD Coordinates
jitter Spatial uncertainty
nxdom,nydom Number of pixel for discretization of the spatial domain in the 
        horizontal and vertical direction respectively. If "auto" is automatically set
        otherwise a numeric vector of length=1
burnin Burnin
POPSMAX Maximum number of pops
POPINIT Initial number of pops
POPMIN Minimum number of pops
ITR Number of iteration
THN Thinning
nullMatrix A matrix with nindiv lines and nloc columns of 0 or 1. For each 
         individual, at each locus it says if the locus is genuinely missing 
         (no attempt to measure it). This info is used under the option 
         filterNA=TRUE do decide how a double missing value should be treated 
         (genuine missing data or double null allele).
}
