% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSample2Map_FUN.R
\name{PlotSample2Map}
\alias{PlotSample2Map}
\title{Plot samples to map colour-coded as for cluster asignment}
\usage{
PlotSample2Map(
  dirIn,
  map = NULL,
  txt,
  pal = "Set1",
  UTM = TRUE,
  w = NULL,
  h = NULL,
  a = 0.2
)
}
\description{
dirIn the directory path where the mcmc results are saved
map a map of class sf
pal palette for sample colours
UTM Whether the coordinats provided in the analysis to Geneland are in UTM (TRUE) 
    or GDS84 / MGA 55 (FALSE)
w, h the extent of the jitter for sample locations to be plotted. NULL (default) if none
a set alpha for transparency
}
